add swagger for bbox country from worldcountriesgenerated
