+++ routes/api_v2.php
@@ -18,6 +18,7 @@ use App\Http\Controllers\V2\CoreTeamLeader\UpdateCoreTeamLeaderController;
 use App\Http\Controllers\V2\Dashboard\ActiveCountriesTableController;
 use App\Http\Controllers\V2\Dashboard\ActiveProjectsTableController;
 use App\Http\Controllers\V2\Dashboard\CountriesController;
+use App\Http\Controllers\V2\Dashboard\CountryDataController;
 use App\Http\Controllers\V2\Dashboard\GetJobsCreatedController;
 use App\Http\Controllers\V2\Dashboard\ProjectListExportController;
 use App\Http\Controllers\V2\Dashboard\GetProjectsController;
@@ -638,7 +639,7 @@ Route::prefix('terrafund')->group(function () {
     Route::put('/site-polygon/{uuid}', [TerrafundEditGeometryController::class, 'updateSitePolygon']);
     Route::post('/site-polygon/{uuid}', [TerrafundEditGeometryController::class, 'createSitePolygon']);
 
-    Route::get('/country/{country}', [TerrafundCreateGeometryController::class, 'getCountryBbox']);
+    
 });
 
 Route::get('/funding-programme', [FundingProgrammeController::class, 'index'])->middleware('i18n');
@@ -671,4 +672,5 @@ Route::prefix('dashboard')->group(function () {
     Route::get('/project-details', ProjectProfileDetailsController::class);
     Route::get('/project-list-export', ProjectListExportController::class);
     Route::get('/get-projects', GetProjectsController::class);
+    Route::get('/country/{country}', [CountryDataController::class, 'getCountryBbox']);
 });
